新增文件下载功能并优化文件上传逻辑
- 新增文件下载接口和前端下载功能
- 优化文件上传逻辑，支持多文件上传
- 调整前端文件操作菜单，增加下载选项
- 修复部分组件类型定义错误- 添加环境变量配置VITE_API_URL

+++ web/src/views/HomeView.vue
@@ -78,9 +78,8 @@ const message = useMessage()
 const dialog = useDialog()
 const options: DropdownOption[] = [
   {
-    label: '编辑',
-    key: 'edit',
-    disabled: true
+    label: '下载',
+    key: 'download'
   },
   {
     label: () => h('span', { style: { color: 'red' } }, '删除'),
@@ -134,21 +133,33 @@ const handleContextmenu = (e: any, file: any) => {
   })
 }
 
-const handleSelect = () => {
+const handleSelect = (key:string) => {
   showDropdown.value = false
-  dialog.success({
-    title: '提示',
-    content: '请再次确认删除该文件？',
-    positiveText: '确定',
-    negativeText: '取消',
-    maskClosable: false,
-    onPositiveClick: () => {
-      deleteFile(selectFile.value.filePath).then((res: any) => {
-        message.success('删除成功')
-        getList()
-      })
-    }
-  })
+  if (key === 'download') {
+    // a 标签
+    const a = document.createElement('a')
+    a.href = '/api/file/download?path=' + selectFile.value.filePath
+    a.download = selectFile.value.fileName
+    a.click()
+    a.remove()
+    return
+  }
+  if (key === 'delete') {
+    dialog.success({
+      title: '提示',
+      content: '请再次确认删除该文件？',
+      positiveText: '确定',
+      negativeText: '取消',
+      maskClosable: false,
+      onPositiveClick: () => {
+        deleteFile(selectFile.value.filePath).then((res: any) => {
+          message.success('删除成功')
+          getList()
+        })
+      }
+    })
+    return
+  }
 }
 const onClickoutside = () => {
   showDropdown.value = false
